Extract preprocess agent
